+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Implemented `Get_BTC_unit_price_in_EUR` test.
+
 ## [0.1.4-preview] - 2025-09-06
 
 - Implemented `GetOfferAsync` method.